Reproduction/requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.19.2
matplotlib==3.3.4
Pillow==8.2.0
scikit-learn==0.24.1
